refactor: 마이페이지 수정
마이페이지 디렉토리 구조 변경
mypage1, mypage2, mypage3를 의미있는 view로 이름 변경
Header를 수정하여 Mypage에서 카테고리 nav바를 제거할 수 있도록 수정
로그인한 사용자만 접근할 수 있도록 수정
회원 정보 수정 view에서 헤더가 나오도록 수정
