+++ src/components/Header.vue
@@ -7,6 +7,14 @@ const activeTab = ref('뮤지컬')
 const setActiveTab = (tab) => {
   activeTab.value = tab
 }
+
+const props = defineProps({
+  isGenre: {
+    type: Boolean,
+    default: true
+  }
+})
+
 </script>
 
 <template>
@@ -15,7 +23,7 @@ const setActiveTab = (tab) => {
     <nav class="navbar bg-body p-2 container-lg">
       <div class="d-flex gap-2 justify-content-between w-100">
         <div class="d-flex w-50">
-          <RouterLink to="/" class="navbar-brand">Picket</RouterLink>
+          <RouterLink to="/" class="navbar-brand"><strong>Pick</strong>et</RouterLink>
           <form class="d-flex w-100" role="search">
             <input class="form-control me-2 border border-secondary" type="search" placeholder="공연을 검색하세요."
               aria-label="Search" />
@@ -24,14 +32,15 @@ const setActiveTab = (tab) => {
         </div>
         <div class="d-flex gap-2 align-items-center">
           <RouterLink class="link-underline-light link-dark" to="/login">로그인</RouterLink>
-          <RouterLink class="link-underline-light link-dark" to="/mypage1">마이페이지</RouterLink>
+          <RouterLink class="link-underline-light link-dark" to="/mypage">마이페이지</RouterLink>
           <RouterLink class="link-underline-light link-dark" to="/seller">판매자 관리 페이지</RouterLink>
         </div>
       </div>
     </nav>
 
     <!-- tab start -->
-    <ul class="nav nav-underline justify-content-center fs-4 border-bottom border-top bg-light-subtle">
+    <ul class="nav nav-underline justify-content-center fs-4 border-bottom border-top bg-light-subtle"
+      v-if="props.isGenre">
       <li class="nav-item" v-for="tab in tabs" :key="tab">
         <a class="nav-link text-black" :class="{ active: activeTab === tab }" href="#"
           @click.prevent="setActiveTab(tab)">
