+++ src/views/mypage/ReviewView.vue
@@ -1,0 +1,88 @@
+<script setup>
+
+</script>
+
+<template>
+    <main class="col-md-9">
+        <h4 class="mb-4">리뷰 및 후기 내역 확인</h4>
+
+        <!-- Date Filter -->
+        <div class="mb-3 d-flex align-items-center gap-2">
+            <span>조회기간:</span>
+            <input type="date" class="form-control form-control-sm" style="max-width: 150px;" />
+            <span>~</span>
+            <input type="date" class="form-control form-control-sm" style="max-width: 150px;" />
+            <button class="btn btn-outline-primary btn-sm">조회</button>
+        </div>
+
+        <!-- Review Table -->
+        <div class="table-responsive">
+            <strong class="--bs-body-color p-2 mb-3 d-inline-block">* 리뷰 내역</strong>
+            <table class="table table-bordered text-center">
+                <thead class="table-light">
+                    <tr>
+                        <th>작성일시</th>
+                        <th>공연명</th>
+                        <th>관람일</th>
+                        <th>리뷰 제목</th>
+                        <th>리뷰 내용</th>
+                        <th>별점</th>
+                        <th>공개 여부</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    <tr v-for="(review, index) in reviews" :key="index">
+                        <td><small>{{ review.date }}</small></td>
+                        <td><small>{{ review.show }}</small></td>
+                        <td><small>{{ review.viewDate }}</small></td>
+                        <td><small>{{ review.title }}</small></td>
+                        <td><small>{{ review.content }}</small></td>
+                        <td>
+                            <i v-for="i in 5" :key="i"
+                                :class="['fa', i <= review.stars ? 'fa-star text-warning' : 'fa-star-o text-secondary']"></i>
+                        </td>
+                        <td><small>{{ review.public }}</small></td>
+                    </tr>
+                </tbody>
+            </table>
+        </div>
+
+        <!-- QnA Table -->
+        <div style="margin-top: 50px;"></div>
+        <div class="table-responsive">
+            <strong class="--bs-body-color p-2 mb-3 d-inline-block">* QnA 내역</strong>
+            <table class="table table-bordered text-center">
+                <thead class="table-light">
+                    <tr>
+                        <th>작성일시</th>
+                        <th>공연명</th>
+                        <th>문의 제목</th>
+                        <th>문의 내용</th>
+                        <th>공개 여부</th>
+                        <th>답변 여부</th>
+                        <th>답변 보기</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    <tr v-for="(qna, index) in qnas" :key="index">
+                        <td><small>{{ qna.date }}</small></td>
+                        <td><small>{{ qna.show }}</small></td>
+                        <td><small>{{ qna.title }}</small></td>
+                        <td><small>{{ qna.content }}</small></td>
+                        <td><small>{{ qna.public }}</small></td>
+                        <td><small>{{ qna.answered }}</small></td>
+                        <td><button class="btn btn-outline-dark btn-sm">답변 확인</button></td>
+                    </tr>
+                </tbody>
+            </table>
+        </div>
+
+        <div class="alert alert-light mt-3 small">
+            <strong>※ 유의사항</strong><br />
+            - 취소 시 예매수수료는 예매일 당일 밤 12시 이전까지 환불됩니다. 그 이후에는 환불되지 않습니다.<br />
+            - 패키지 구성 상품별로 회차와 좌석지정을 해야 공연 관람이 가능합니다.
+        </div>
+    </main>
+</template>
+
+<style scoped></style>
